Making all associations explicit

+++ LAB1/dst1_0626885/src/dst1/model/EnvironmentDAO.java
@@ -30,21 +30,28 @@ public class EnvironmentDAO {
 	
 	/**
 	 * Removes an environment and its related jobs
-	 * @param environment the environment to remove
+	 * @param environmentId the environmentId of the environment to remove
+	 * @return true if removed successfully, false otherwise
 	 */
 	@SuppressWarnings("unchecked")
-	public void removeEnvironment(Environment environment) {
+	public boolean removeEnvironment(Long environmentId) {
+		
+		Environment environment_ = entityManager.find(Environment.class, environmentId);
+		
+		if(environment_ == null)
+			return false;
 
 		List<Job> jobResult = entityManager.createQuery(
 										"select j from Job j where j.environment.environmentId = :id"
-								).setParameter("id", environment.getEnvironmentId())
+								).setParameter("id", environment_.getEnvironmentId())
 								 .getResultList();
 		
 		for(Job job : jobResult) 
 			entityManager.remove(job);
 			
-		entityManager.remove(environment);
-				
+		entityManager.remove(environment_);
+		
+		return true;
 	}
 	
 }
